Clarify code and doc related to SPIRVAddRequirementsPass
The SPIRVAddRequirements inserts the appropriate OpCapability and
OpExtension instructions. The code and documentation are clarified in
that respect. A test case is also added, showing that the backend
support the SPV_KHR_no_integer_wrap_decoration extension.

The code is also simplified in a few places to use references when
pointers are never null and therefore not needed.