CUDA/HIP: Allow __int128 on the host side
Consider case where __int128 type is supported by the host target but
not by a device target (e.g. spirv*). Clang emits an error message for
unsupported type even if the device code does not use it. This patch
fixes this issue by emitting the error message when the device code
attempts to use the unsupported type.

Differential Revision: https://reviews.llvm.org/D111047