[SPIR-V] Fix conversion errors when building with msvc
SPIR-V symbolic operand enums such as LinkageType, Decoration etc. are
generated using TableGen as int32_t values and currently there is no
option to change this. In order to silence warnings and eliminate
errors when building with the msvc compiler, the values need to be
statically casted to uint32_t.