Enable topological sort of the module's global entities, i.e. types/c… (#166)
* Enable topological sort of the module's global entities, i.e. types/consts/etc.

In general this is needed to ensure proper ordering which was previously ensured by using MapVector
and hoping everything's added to it in the correct order (not always true). While MapVector stays to make sorting stable
to avoid related modifications of LIT tests, now the proper declaration order is ensured according to the spec.